+++ db/migrate/20201211203158_create_table.rb
@@ -1,7 +1,7 @@
 class CreateTable < ActiveRecord::Migration[6.0]
   def change
 	 create_table :tables do |t|
-		t.integer :app_id
+		t.bigint :app_id
 		t.string :name
 		t.timestamps
     end
